added pages for sprite edit dialog with initial resolution resizing window
