+++ editor/src/com/talosvfx/talos/editor/addons/scene/events/AssetResolutionChanged.java
@@ -1,0 +1,10 @@
+package com.talosvfx.talos.editor.addons.scene.events;
+
+import com.talosvfx.talos.editor.notifications.TalosEvent;
+
+public class AssetResolutionChanged implements TalosEvent {
+    @Override
+    public void reset() {
+
+    }
+}
